backend/requirements.txt
# QuantEdge Backend Dependencies
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0

# Database
sqlalchemy==2.0.23
aiosqlite==0.19.0

# Data Processing
pandas==2.1.3
numpy==1.26.2
yfinance==0.2.32
ta==0.11.0  # Technical Analysis library

# Machine Learning
tensorflow==2.15.0
scikit-learn==1.3.2
stable-baselines3==2.2.1

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6
pytz==2024.1

# Logging & Monitoring
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
